9465: Create feature flag for section outbox number of days [skip ci]
